Phase 13-1 (034): KRW int 전환 + ORDER BY 표준화 (13-s·13-t) — 885 PASS / 76%
- 트랙 A: Portfolio/CashManager/ExecutionModel 금액 필드 float→int 전환
  + alembic 마이그레이션 (REAL→INTEGER) + 신규 32건
- 트랙 B: 12개 repository/service 쿼리에 결정론적 ORDER BY 추가 + 신규 9건
- 13번 정확성 정책 문서 20/20 100% 완료 ✅

+++ backend/app/models/backtest.py
@@ -59,7 +59,7 @@ class BacktestRun(Base):
     # 기간 / 초기자금
     start_date: Mapped[date_type] = mapped_column(Date, nullable=False)
     end_date: Mapped[date_type] = mapped_column(Date, nullable=False)
-    initial_cash: Mapped[float] = mapped_column(Float, nullable=False)
+    initial_cash: Mapped[int] = mapped_column(Integer, nullable=False)  # KRW 정수 §14
 
     # 비용 / 정확성 정책 스냅샷
     fee_rate: Mapped[float] = mapped_column(Float, nullable=False)
@@ -116,9 +116,9 @@ class BacktestResult(Base, TimestampMixin):
     )
 
     # 핵심 지표 (calculate_metrics 키와 동일 의미)
-    initial_cash: Mapped[float] = mapped_column(Float, nullable=False)
-    final_equity: Mapped[float] = mapped_column(Float, nullable=False)
-    total_return_pct: Mapped[float] = mapped_column(Float, nullable=False)
+    initial_cash: Mapped[int] = mapped_column(Integer, nullable=False)    # KRW 정수 §14
+    final_equity: Mapped[int] = mapped_column(Integer, nullable=False)    # KRW 정수 §14
+    total_return_pct: Mapped[float] = mapped_column(Float, nullable=False)  # 비율 float
     annual_return_pct: Mapped[float] = mapped_column(Float, default=0.0, nullable=False)
     mdd_pct: Mapped[float] = mapped_column(Float, default=0.0, nullable=False)
